Simple AggregatedFrame PDU compatibility
Is a simple and fast solution to be compatible with AggregatedFrame
PDU's.

The response to any AggregatedFrame PDU will be another AggregatedFrame
PDU with each of the responses to the initial inner PDU's in the same
order.

The solution can be much more depurated... but it works, according to
the NFC Forum LLCP specs.

Tested using Blackberry 7.1 OS.